Move StatusMonitor config files
