+++ hedera-node/hedera-config/src/testFixtures/java/module-info.java
@@ -17,6 +17,7 @@ module com.hedera.node.config.test.fixtures {
     requires org.hiero.consensus.metrics;
     requires org.hiero.consensus.reconnect;
     requires org.hiero.consensus.state;
+    requires org.hiero.consensus.status.monitor;
     requires org.hiero.consensus.utility;
     requires static transitive com.github.spotbugs.annotations;
 }
