+++ platform-sdk/consensus-status-monitor/src/main/java/module-info.java
@@ -17,7 +17,6 @@ module org.hiero.consensus.status.monitor {
     requires org.hiero.base.utility;
     requires org.hiero.consensus.metrics;
     requires org.hiero.consensus.roster;
-    requires org.hiero.consensus.utility;
     requires org.apache.logging.log4j;
     requires static transitive com.github.spotbugs.annotations;
