+++ platform-sdk/consensus-status-monitor/src/main/java/org/hiero/consensus/status/monitor/StatusMonitorModule.java
@@ -13,7 +13,6 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.metrics.api.Metrics;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.time.Duration;
-import org.hiero.consensus.config.PlatformStatusConfig;
 import org.hiero.consensus.model.hashgraph.ConsensusRound;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.model.notification.IssNotification;
@@ -21,6 +20,7 @@ import org.hiero.consensus.model.quiescence.QuiescenceCommand;
 import org.hiero.consensus.model.state.StateSavingResult;
 import org.hiero.consensus.model.status.PlatformStatus;
 import org.hiero.consensus.status.monitor.actions.PlatformStatusAction;
+import org.hiero.consensus.status.monitor.config.PlatformStatusConfig;
 import org.hiero.consensus.status.monitor.config.StatusMonitorWiringConfig;
 import org.hiero.consensus.status.monitor.internal.DefaultPlatformMonitor;
 import org.hiero.consensus.status.monitor.internal.PlatformMonitor;
