+++ platform-sdk/consensus-status-monitor/src/main/java/org/hiero/consensus/status/monitor/config/UptimeConfig.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.config;
+package org.hiero.consensus.status.monitor.config;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
