+++ platform-sdk/consensus-status-monitor/src/main/java/org/hiero/consensus/status/monitor/logic/ReplayingEventsStatusLogic.java
@@ -3,10 +3,10 @@ package org.hiero.consensus.status.monitor.logic;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
-import org.hiero.consensus.config.PlatformStatusConfig;
 import org.hiero.consensus.model.status.PlatformStatus;
 import org.hiero.consensus.status.monitor.actions.DoneReplayingEventsAction;
 import org.hiero.consensus.status.monitor.actions.FreezePeriodEnteredAction;
+import org.hiero.consensus.status.monitor.config.PlatformStatusConfig;
 
 /**
  * Class containing the state machine logic for the {@link PlatformStatus#REPLAYING_EVENTS} status.
