+++ platform-sdk/consensus-status-monitor/src/test/java/org/hiero/consensus/status/monitor/logic/ActiveStatusLogicTests.java
@@ -11,8 +11,6 @@ import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import org.hiero.consensus.config.PlatformStatusConfig;
-import org.hiero.consensus.config.PlatformStatusConfig_;
 import org.hiero.consensus.model.status.PlatformStatus;
 import org.hiero.consensus.status.monitor.actions.CatastrophicFailureAction;
 import org.hiero.consensus.status.monitor.actions.DoneReplayingEventsAction;
@@ -24,6 +22,8 @@ import org.hiero.consensus.status.monitor.actions.StartedReplayingEventsAction;
 import org.hiero.consensus.status.monitor.actions.StateWrittenToDiskAction;
 import org.hiero.consensus.status.monitor.actions.TimeElapsedAction;
 import org.hiero.consensus.status.monitor.actions.TimeElapsedAction.QuiescingStatus;
+import org.hiero.consensus.status.monitor.config.PlatformStatusConfig;
+import org.hiero.consensus.status.monitor.config.PlatformStatusConfig_;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
