+++ platform-sdk/consensus-status-monitor/src/test/java/org/hiero/consensus/status/monitor/logic/ReplayingEventsStatusLogicTests.java
@@ -9,7 +9,6 @@ import com.swirlds.base.test.fixtures.time.FakeTime;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
 import java.time.temporal.ChronoUnit;
-import org.hiero.consensus.config.PlatformStatusConfig;
 import org.hiero.consensus.model.status.PlatformStatus;
 import org.hiero.consensus.status.monitor.actions.CatastrophicFailureAction;
 import org.hiero.consensus.status.monitor.actions.DoneReplayingEventsAction;
@@ -20,6 +19,7 @@ import org.hiero.consensus.status.monitor.actions.SelfEventReachedConsensusActio
 import org.hiero.consensus.status.monitor.actions.StartedReplayingEventsAction;
 import org.hiero.consensus.status.monitor.actions.StateWrittenToDiskAction;
 import org.hiero.consensus.status.monitor.actions.TimeElapsedAction;
+import org.hiero.consensus.status.monitor.config.PlatformStatusConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
